30 jun

+++ institucional/aluno/index.php
@@ -2,6 +2,7 @@
     require_once "../../util/config.php";
     session_start();
 
+    if ($_SESSION != null){
     $idAdmin = $_SESSION['idAdmin'];
     $sql = "SELECT * FROM aluno";
     $result = mysqli_query($link, $sql);
@@ -63,6 +64,15 @@
     <title>Alunos</title>
     <link rel="stylesheet" href="../../css/mural.css">
     <link rel="icon" href="../../imagens/nucleo-adra-icone.png" >
+    <!-- Confirmação de Exclusão -->
+    <script>
+        function confirmarExclusao(event) {
+            var confirmacao = confirm("Você tem certeza que quer excluir?");
+            if (!confirmacao) {
+                event.preventDefault(); // Cancela a ação de exclusão se o usuário clicar em "Não"
+            }
+        }
+    </script>
 </head>
 <body>
 <header>
@@ -152,7 +162,7 @@
             ?></td>
             <td><?php echo('<a href="read.php?id='.$row['id'].'" class="crud_link">Ver mais</a>')?></td>
             <td><?php echo('<a href="update.php?id='.$row['id'].'" class="crud_link">Alterar</a>')?></td>
-            <td><?php echo('<a href="delete.php?id='.$row['id'].'" class="crud_link">Excluir</a>')?></td>
+            <td><?php echo('<a href="delete.php?id='.$row['id'].'" class="crud_link" onclick="confirmarExclusao(event)">Excluir</a>')?></td>
         </tr>
         <?php } ?>
     </table>
@@ -161,6 +171,11 @@
     </div>
     </div>
 </div>
-
+<?php
+    } else{
+    // Redirecionamento de volta para a página anterior
+    header("Location: ../../index.php");
+    exit(); // Certifique-se de sair após o redirecionamento
+    }?>
 </body>
 </html>
